+++ src/sys/os/linux/syscalls/aarch64.rs
@@ -8,14 +8,14 @@
 use super::{LinuxOffset, shared_docs::*};
 use crate::{
     AT_FDCWD, LINUX_SYS as SYS, Linux, LinuxSigaction, LinuxStat, LinuxTimespec, asm, c_char,
-    c_int, c_uint, c_ulong,
+    c_int, c_uchar, c_uint, c_ulong,
 };
 
 /// # Syscalls: File descriptors.
 impl Linux {
     #[must_use]
     #[doc = SYS_READ!()]
-    pub unsafe fn sys_read(fd: c_int, buf: *mut u8, count: usize) -> isize {
+    pub unsafe fn sys_read(fd: c_int, buf: *mut c_uchar, count: usize) -> isize {
         let result;
         unsafe {
             asm!(
@@ -33,7 +33,7 @@ impl Linux {
     }
     #[must_use]
     #[doc = SYS_WRITE!()]
-    pub unsafe fn sys_write(fd: c_int, buf: *const u8, count: usize) -> isize {
+    pub unsafe fn sys_write(fd: c_int, buf: *const c_uchar, count: usize) -> isize {
         let result;
         unsafe {
             asm!(
@@ -197,7 +197,7 @@ impl Linux {
 
     #[must_use]
     #[doc = SYS_GETDENTS!()]
-    pub unsafe fn sys_getdents(fd: c_int, dirp: *mut u8, count: usize) -> isize {
+    pub unsafe fn sys_getdents(fd: c_int, dirp: *mut c_uchar, count: usize) -> isize {
         let result: isize;
         unsafe {
             asm!(
@@ -219,7 +219,7 @@ impl Linux {
 impl Linux {
     #[must_use]
     #[doc = SYS_IOCTL!()]
-    pub unsafe fn sys_ioctl(fd: c_int, request: c_ulong, argp: *mut u8) -> isize {
+    pub unsafe fn sys_ioctl(fd: c_int, request: c_ulong, argp: *mut c_uchar) -> isize {
         let result;
         unsafe {
             asm!(
@@ -291,7 +291,7 @@ impl Linux {
     }
     #[must_use]
     #[doc = SYS_GETPID!()]
-    pub unsafe fn sys_getpid() -> i32 {
+    pub unsafe fn sys_getpid() -> c_int {
         let result: isize;
         unsafe {
             asm!(
@@ -302,11 +302,11 @@ impl Linux {
                 options(nostack)
             );
         }
-        result as i32
+        result as c_int
     }
     #[must_use]
     #[doc = SYS_GETRANDOM!()]
-    pub unsafe fn sys_getrandom(buffer: *mut u8, size: usize, flags: c_uint) -> isize {
+    pub unsafe fn sys_getrandom(buffer: *mut c_uchar, size: usize, flags: c_uint) -> isize {
         let result;
         unsafe {
             asm!(
@@ -328,6 +328,24 @@ impl Linux {
 /// # Syscalls: Timing and signal handling.
 impl Linux {
     #[must_use]
+    #[doc = SYS_CLOCK_GETRES!()]
+    pub unsafe fn sys_clock_getres(clock_id: c_int, res: *mut LinuxTimespec) -> isize {
+        let result;
+        unsafe {
+            asm!(
+                "mov x8, {CLOCK_GETRES}",
+                "svc 0",
+                CLOCK_GETRES = const SYS::CLOCK_GETRES,
+                in("x0") clock_id,
+                in("x1") res,
+                lateout("x0") result,
+                options(nostack)
+            );
+        }
+        result
+    }
+
+    #[must_use]
     #[doc = SYS_CLOCK_GETTIME!()]
     pub unsafe fn sys_clock_gettime(clock_id: c_int, tp: *mut LinuxTimespec) -> isize {
         let result;
@@ -341,8 +359,8 @@ impl Linux {
                 lateout("x0") result,
                 options(nostack)
             );
-            result
         }
+        result
     }
     #[must_use]
     #[doc = SYS_NANOSLEEP!()]