+++ src/sys/os/linux/syscalls/x86_64.rs
@@ -10,15 +10,15 @@
 
 use super::{LinuxOffset, shared_docs::*};
 use crate::{
-    LINUX_SYS as SYS, Linux, LinuxSigaction, LinuxStat, LinuxTimespec, asm, c_char, c_int, c_uint,
-    c_ulong,
+    LINUX_SYS as SYS, Linux, LinuxSigaction, LinuxStat, LinuxTimespec, asm, c_char, c_int, c_uchar,
+    c_uint, c_ulong,
 };
 
 /// # Syscalls: File descriptors.
 impl Linux {
     #[must_use]
     #[doc = SYS_READ!()]
-    pub unsafe fn sys_read(fd: c_int, buf: *mut u8, count: usize) -> isize {
+    pub unsafe fn sys_read(fd: c_int, buf: *mut c_uchar, count: usize) -> isize {
         let result;
         unsafe {
             asm!(
@@ -38,7 +38,7 @@ impl Linux {
     }
     #[must_use]
     #[doc = SYS_WRITE!()]
-    pub unsafe fn sys_write(fd: c_int, buf: *const u8, count: usize) -> isize {
+    pub unsafe fn sys_write(fd: c_int, buf: *const c_uchar, count: usize) -> isize {
         let result;
         unsafe {
             asm!(
@@ -199,7 +199,7 @@ impl Linux {
     }
     #[must_use]
     #[doc = SYS_GETDENTS!()]
-    pub unsafe fn sys_getdents(fd: c_int, dirp: *mut u8, count: usize) -> isize {
+    pub unsafe fn sys_getdents(fd: c_int, dirp: *mut c_uchar, count: usize) -> isize {
         let result;
         unsafe {
             asm!(
@@ -225,7 +225,7 @@ impl Linux {
 impl Linux {
     #[must_use]
     #[doc = SYS_IOCTL!()]
-    pub unsafe fn sys_ioctl(fd: c_int, request: c_ulong, argp: *mut u8) -> isize {
+    pub unsafe fn sys_ioctl(fd: c_int, request: c_ulong, argp: *mut c_uchar) -> isize {
         let result;
         unsafe {
             asm!(
@@ -303,7 +303,7 @@ impl Linux {
     }
     #[must_use]
     #[doc = SYS_GETPID!()]
-    pub unsafe fn sys_getpid() -> i32 {
+    pub unsafe fn sys_getpid() -> c_int {
         let result: isize;
         unsafe {
             asm!(
@@ -314,11 +314,11 @@ impl Linux {
                 options(nostack)
             );
         }
-        result as i32
+        result as c_int
     }
     #[must_use]
     #[doc = SYS_GETRANDOM!()]
-    pub unsafe fn sys_getrandom(buffer: *mut u8, size: usize, flags: c_uint) -> isize {
+    pub unsafe fn sys_getrandom(buffer: *mut c_uchar, size: usize, flags: c_uint) -> isize {
         let result;
         unsafe {
             asm!(
@@ -341,6 +341,25 @@ impl Linux {
 /// # Syscalls: Timing and signal handling.
 impl Linux {
     #[must_use]
+    #[doc = SYS_CLOCK_GETRES!()]
+    pub unsafe fn sys_clock_getres(clock_id: c_int, res: *mut LinuxTimespec) -> isize {
+        let result;
+        unsafe {
+            asm!(
+                "mov rax, {CLOCK_GETRES}",
+                "syscall",
+                CLOCK_GETRES = const SYS::CLOCK_GETRES,
+                in("rdi") clock_id,
+                in("rsi") res,
+                lateout("rax") result,
+                lateout("rcx") _,
+                lateout("r11") _,
+                options(nostack)
+            );
+        }
+        result
+    }
+    #[must_use]
     #[doc = SYS_CLOCK_GETTIME!()]
     pub unsafe fn sys_clock_gettime(clock_id: c_int, tp: *mut LinuxTimespec) -> isize {
         let result;
@@ -356,8 +375,8 @@ impl Linux {
                 lateout("r11") _,
                 options(nostack)
             );
-            result
         }
+        result
     }
     #[must_use]
     #[doc = SYS_NANOSLEEP!()]